# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Habit {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface HabitWithStatus {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  completedToday: boolean;
  streak: number;
}

export interface HabitInput {
  /** @minLength 1 */
  title: string;
  description?: string;
}

export interface HabitUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
}

export interface CompletionInput {
  habitId: number;
  /** YYYY-MM-DD */
  date: string;
}

export interface CompletionResult {
  completed: boolean;
  habitId: number;
  date: string;
}

export interface HabitHistory {
  habitId: number;
  completedDates: string[];
}

export interface WeeklyBar {
  /** YYYY-MM-DD week start */
  date: string;
  count: number;
}

export interface HabitStreakItem {
  habitId: number;
  title: string;
  streak: number;
}

export interface DashboardStats {
  completedToday: number;
  totalHabits: number;
  completedThisWeek: number;
  weeklyBars: WeeklyBar[];
  habitStreaks: HabitStreakItem[];
}