# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Habit Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: habits
    description: Habit management
  - name: tracking
    description: Habit tracking and history
  - name: dashboard
    description: Dashboard aggregates
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /habits:
    get:
      operationId: getHabits
      tags: [habits]
      summary: List habits with today's status and streak
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitWithStatus"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a new habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Habit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitWithStatus"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /habits/{id}:
    put:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitWithStatus"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit and its entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /complete:
    post:
      operationId: toggleCompletion
      tags: [tracking]
      summary: Toggle today's completion for a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionInput"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /history/{habitId}:
    get:
      operationId: getHistory
      tags: [tracking]
      summary: Get completion history for a habit (last 30 days)
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completion history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitHistory"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard aggregate stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    Habit:
      type: object
      required: [id, userId, title, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    HabitWithStatus:
      type: object
      required: [id, userId, title, createdAt, completedToday, streak]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedToday:
          type: boolean
        streak:
          type: integer

    HabitInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string

    HabitUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]

    CompletionInput:
      type: object
      required: [habitId, date]
      properties:
        habitId:
          type: integer
        date:
          type: string
          description: "YYYY-MM-DD"

    CompletionResult:
      type: object
      required: [completed, habitId, date]
      properties:
        completed:
          type: boolean
        habitId:
          type: integer
        date:
          type: string

    HabitHistory:
      type: object
      required: [habitId, completedDates]
      properties:
        habitId:
          type: integer
        completedDates:
          type: array
          items:
            type: string
            description: "YYYY-MM-DD"

    WeeklyBar:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
          description: "YYYY-MM-DD week start"
        count:
          type: integer

    DashboardStats:
      type: object
      required: [completedToday, totalHabits, completedThisWeek, weeklyBars, habitStreaks]
      properties:
        completedToday:
          type: integer
        totalHabits:
          type: integer
        completedThisWeek:
          type: integer
        weeklyBars:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBar"
        habitStreaks:
          type: array
          items:
            $ref: "#/components/schemas/HabitStreakItem"

    HabitStreakItem:
      type: object
      required: [habitId, title, streak]
      properties:
        habitId:
          type: integer
        title:
          type: string
        streak:
          type: integer