# lib/api-zod/src/generated/types/completionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletionInput {
  habitId: number;
  /** YYYY-MM-DD */
  date: string;
}