# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
}