# lib/api-zod/src/generated/types/habitHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitHistory {
  habitId: number;
  completedDates: string[];
}