# lib/api-zod/src/generated/types/habitStreakItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitStreakItem {
  habitId: number;
  title: string;
  streak: number;
}