# lib/api-zod/src/generated/types/habitWithStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitWithStatus {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
  completedToday: boolean;
  streak: number;
}